requirements.txt
fastapi==0.128.6
uvicorn==0.40.0
pydantic==2.12.5
python-dotenv==1.2.1
google-genai==1.62.0
chromadb==1.5.0
python-multipart==0.0.22
pydantic-settings==2.12.0
twilio==9.10.1
